- replaced PhpTemplates by ModuleTemplates - fixed install and uninstall

+++ modules/CLLIBR/setup/uninstall.php
@@ -2,25 +2,30 @@
 /**
  * Online library for Claroline
  *
- * @version     CLLIBR 0.3.0 $Revision$ - Claroline 1.9
+ * @version     CLLIBR 0.5.0 $Revision$ - Claroline 1.9
  * @copyright   2001-2011 Universite catholique de Louvain (UCL)
  * @license     http://www.gnu.org/copyleft/gpl.html (GPL) GENERAL PUBLIC LICENSE
  * @package     CLLIBR
  * @author      Frederic Fervaille <frederic.fervaille@uclouvain.be>
  */
 
-//$cllibr_path = get_path( 'rootSys' ) . 'cllibrary';
-$cllibr_path = get_conf( 'CLLIBR_storage_directory' );
+load_module_config( 'CLLIBR' );
 
-$cllibr_dir = new DirectoryIterator( $cllibr_path );
+$cllibr_path = get_conf( 'CLLIBR_storage_directory' )
+             ? get_conf( 'CLLIBR_storage_directory' )
+             : get_path( 'rootSys' ) . 'cllibrary';
 
-foreach ( $cllibr_dir as $file )
+if ( file_exists( $cllibr_path ) )
 {
-    if ( ! $file->isDot() )
+    $cllibr_dir = new DirectoryIterator( $cllibr_path );
+    
+    foreach ( $cllibr_dir as $file )
     {
-        unlink( $cllibr_path . $file->getFileName() );
+        if ( ! $file->isDot() )
+        {
+            unlink( $cllibr_path . $file->getFileName() );
+        }
     }
+    
+    rmdir( $cllibr_path );
 }
-
-rmdir( $cllibr_path );
-
